requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-community
langchain-ollama
sqlalchemy
pymysql
python-dotenv
bcrypt
